refactor(windows): split installer_windows.go by concern (#110)
installer_windows.go was the last undifferentiated slab of the Windows
platform layer: 794 lines and ~40 functions covering system probing,
BitLocker, partitioning, artifact download, uninstall and the shared
exec helpers, with no boundary between any of them.

This is a pure code move. Every function body and every comment is
carried over verbatim — including the long rationale comments that
encode run IDs and issue numbers (the #63 PendingFileRenameOperations
story, the 8136ae6 VHDX→raw switch, the #53 fail-closed checksum rule).
The only added lines are three `// ── … ──` section headers and the
per-file build tag and import blocks. No identifier, signature, or
statement changed.

The split follows what the code actually clusters around:

  installer_windows.go   220  install/uninstall lifecycle: getSystemInfo,
                              validatePlatformConfig, checkSystem,
                              createDirectories, uninstall{,With},
                              getUninstallInfo, rebootWindows
  disk_windows.go        178  volumes and partitions: defrag,
                              listDataPartitions, dedicatedVolumeInfo,
                              CreateDataPartition, removePartitionAndExtendC,
                              createRootDisk (+ VDL extension)
  sysprobe_windows.go    167  firmware/platform/power probes: isAdmin,
                              isUEFI, Secure Boot, Fast Startup,
                              battery, pending reboot, hibernation, RAM
  deployer_windows.go    129  boot-artifact acquisition and the
                              fail-closed SHA256SUMS verification (#53)
  bitlocker_windows.go    99  encryption state and recovery-key capture
  winexec_windows.go      56  the shared Windows substrate: runCmd,
                              runPowerShell{,Output}, restrictFileACL,
                              storageDrive/wootcDir

Two deviations from the plan in #110. There is no esp_windows.go or
bcd_windows.go: the ESP and BCD concerns were already extracted to
installer_esp.go in 790c66a and 8386215, and nothing ESP- or BCD-shaped
remained here beyond uninstallWith's cleanup call, which belongs to the
uninstall flow. In their place, two clusters the issue did not
anticipate earned files of their own — the deployer download with its
checksum verification, and the exec/path helpers, which are the
substrate seven other *_windows.go files already build on rather than
anything installer-specific.

Verified: windows/amd64 build (desktop,production,native_webview2loader),
linux `go build ./...`, `go test ./...`, and windows/amd64 `go vet` all
pass. A line-level diff confirms the union of the six files is the
original file plus the three headers.
